+++ include/entropy.hpp
@@ -3,7 +3,7 @@
  */
 
 #include "blockmodel/blockmodel.hpp"
-#include "blockmodel/dist_blockmodel.hpp"
+#include "distributed/two_hop_blockmodel.hpp"
 #include "common.hpp"
 #include "delta.hpp"
 #include "utils.hpp"
@@ -11,30 +11,30 @@
 namespace entropy {
 
 /// Computes the change in entropy under a proposed block merge.
-double block_merge_delta_mdl(int current_block, int proposal, int num_edges, const Blockmodel &blockmodel,
+double block_merge_delta_mdl(long current_block, long proposal, long num_edges, const Blockmodel &blockmodel,
                              EdgeCountUpdates &updates, common::NewBlockDegrees &block_degrees);
 
 /// Computes the change in entropy under a proposed block merge using sparse intermediate structures.
-double block_merge_delta_mdl(int current_block, int proposal, int num_edges, const Blockmodel &blockmodel,
+double block_merge_delta_mdl(long current_block, long proposal, long num_edges, const Blockmodel &blockmodel,
                              SparseEdgeCountUpdates &updates, common::NewBlockDegrees &block_degrees);
 
 /// Computes the change in entropy under a proposed block merge using changes to the blockmodel.
-double block_merge_delta_mdl(int current_block, const Blockmodel &blockmodel, const Delta &delta,
+double block_merge_delta_mdl(long current_block, const Blockmodel &blockmodel, const Delta &delta,
                              common::NewBlockDegrees &block_degrees);
 
 /// Computes the change in entropy under a proposed block merge using changes to the blockmodel. This method should
 /// be preferred in almost all cases.
-double block_merge_delta_mdl(int current_block, utils::ProposalAndEdgeCounts proposal, const Blockmodel &blockmodel,
+double block_merge_delta_mdl(long current_block, utils::ProposalAndEdgeCounts proposal, const Blockmodel &blockmodel,
                              const Delta &delta);
 
 /// Computes the change in blockmodel minimum description length when a vertex moves from `current_block` to `proposal`.
 /// Uses a dense version of `updates` to the blockmodel, and requires pre-calculated updated `block_degrees`.
-double delta_mdl(int current_block, int proposal, const Blockmodel &blockmodel, int num_edges,
+double delta_mdl(long current_block, long proposal, const Blockmodel &blockmodel, long num_edges,
                  EdgeCountUpdates &updates, common::NewBlockDegrees &block_degrees);
 
 /// Computes the change in blockmodel minimum description length when a vertex moves from `current_block` to `proposal`.
 /// Uses a sparse version of `updates` to the blockmodel, and requires pre-calculated updated `block_degrees`.
-double delta_mdl(int current_block, int proposal, const Blockmodel &blockmodel, int num_edges,
+double delta_mdl(long current_block, long proposal, const Blockmodel &blockmodel, long num_edges,
                  SparseEdgeCountUpdates &updates, common::NewBlockDegrees &block_degrees);
 
 /// Computes the change in blockmodel minimum description length when a vertex moves from one block to another. Uses
@@ -53,12 +53,12 @@ double hastings_correction(const Blockmodel &blockmodel, EdgeWeights &out_blocks
                            common::NewBlockDegrees &new_block_degrees);
 
 /// Computes the hastings correction using the blockmodel deltas under the proposed vertex move.
-double hastings_correction(int vertex, const Graph &graph, const Blockmodel &blockmodel, const Delta &delta,
-                           int current_block, const utils::ProposalAndEdgeCounts &proposal);
+double hastings_correction(long vertex, const Graph &graph, const Blockmodel &blockmodel, const Delta &delta,
+                           long current_block, const utils::ProposalAndEdgeCounts &proposal);
 
 /// Calculates the minimum description length of `blockmodel` for a directed graph with `num_vertices` vertices and
 /// `num_edges` edges.
-double mdl(const Blockmodel &blockmodel, int num_vertices, int num_edges);
+double mdl(const Blockmodel &blockmodel, long num_vertices, long num_edges);
 
 /// Computes the normalized minimum description length using `null_mdl_v1`.
 double normalize_mdl_v1(double mdl, long num_edges);
@@ -78,11 +78,11 @@ double null_mdl_v2(long num_vertices, long num_edges);
 namespace dist {
 
 /// Computes the overall entropy of the given blockmodel for a directed graph.
-double mdl(const TwoHopBlockmodel &blockmodel, int num_vertices, int num_edges);
+double mdl(const TwoHopBlockmodel &blockmodel, long num_vertices, long num_edges);
 
 // TODO: add an undirected distributed mdl
 
-// TODO: handle case when number of edges/vertices >= 2.15 billion (int --> long)
+// TODO: handle case when number of edges/vertices >= 2.15 billion (long --> long)
 
 }
