+++ include/partition.hpp
@@ -13,15 +13,15 @@
 namespace partition {
 
 /// Partitions the graph using the method chosen via `args`
-Graph partition(const Graph &graph, int rank, int num_processes, Args &args);
+Graph partition(const Graph &graph, long rank, long num_processes, Args &args);
 
 /// Partitions the graph using the round robin strategy. The resulting partitions do not overlap.
-Graph partition_round_robin(const Graph &graph, int rank, int num_processes, int target_num_vertices);
+Graph partition_round_robin(const Graph &graph, long rank, long num_processes, long target_num_vertices);
 
 /// Partitions the graph using the random strategy. The resulting partitions do not overlap.
-Graph partition_random(const Graph &graph, int rank, int num_processes, int target_num_vertices);
+Graph partition_random(const Graph &graph, long rank, long num_processes, long target_num_vertices);
 
 /// Partitions the graph using the snowball sampling strategy. The resulting partitions DO overlap.
-Graph partition_snowball(const Graph &graph, int rank, int num_processes, int target_num_vertices);
+Graph partition_snowball(const Graph &graph, long rank, long num_processes, long target_num_vertices);
 
 }
